refactor: use PartitionDistributorConfig.ZoneSpec as the single ZoneSpec type
